COMMERCE-11740 convert quantity to BigDecimal

+++ modules/apps/commerce/commerce-inventory-api/src/main/java/com/liferay/commerce/inventory/service/CommerceInventoryBookedQuantityLocalService.java
@@ -28,6 +28,7 @@ import com.liferay.portal.kernel.util.OrderByComparator;
 
 import java.io.Serializable;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
@@ -59,7 +60,7 @@ public interface CommerceInventoryBookedQuantityLocalService
 	 */
 	@Indexable(type = IndexableType.REINDEX)
 	public CommerceInventoryBookedQuantity addCommerceBookedQuantity(
-			long userId, Date expirationDate, int quantity, String sku,
+			long userId, Date expirationDate, BigDecimal quantity, String sku,
 			String unitOfMeasureKey, Map<String, String> context)
 		throws PortalException;
 
@@ -80,7 +81,7 @@ public interface CommerceInventoryBookedQuantityLocalService
 	public void checkCommerceInventoryBookedQuantities();
 
 	public CommerceInventoryBookedQuantity consumeCommerceBookedQuantity(
-			long commerceBookedQuantityId, int quantity)
+			long commerceBookedQuantityId, BigDecimal quantity)
 		throws NoSuchInventoryBookedQuantityException;
 
 	/**
@@ -300,7 +301,7 @@ public interface CommerceInventoryBookedQuantityLocalService
 
 	public CommerceInventoryBookedQuantity resetCommerceBookedQuantity(
 			long commerceBookedQuantityId, long userId, Date expirationDate,
-			int quantity, String sku, Map<String, String> context)
+			BigDecimal quantity, String sku, Map<String, String> context)
 		throws PortalException;
 
 	public CommerceInventoryBookedQuantity
@@ -337,8 +338,8 @@ public interface CommerceInventoryBookedQuantityLocalService
 	@Indexable(type = IndexableType.REINDEX)
 	public CommerceInventoryBookedQuantity
 			updateCommerceInventoryBookedQuantity(
-				long userId, long commerceInventoryBookedQuantityId,
-				int quantity, Map<String, String> context, long mvccVersion)
+			long userId, long commerceInventoryBookedQuantityId,
+			BigDecimal quantity, Map<String, String> context, long mvccVersion)
 		throws PortalException;
 
 }
